feat: refactor routing and implement barrel style imports for features directory
