+++ frontend/src/features/Dashboard/index.ts
@@ -1,0 +1,3 @@
+import * as pages from './pages';
+
+export { pages };
