+++ frontend/src/features/Dashboard/pages/index.ts
@@ -1,0 +1,1 @@
+export { Dashboard } from './Dashboard/Dashboard';
