+++ frontend/src/features/Login/index.ts
@@ -1,0 +1,3 @@
+import * as pages from './pages';
+
+export { pages };
