+++ frontend/src/features/index.ts
@@ -1,0 +1,4 @@
+import * as dashboard from './dashboard';
+import * as login from './login';
+
+export { dashboard, login };
